Fix bug when creating and updating project
