--- src/main/java/ues/occ/proyeccion/social/ws/app/dto/ProyectoEstudianteDto.java
@@ -1,8 +1,0 @@
-package ues.occ.proyeccion.social.ws.app.dto;
-
-import lombok.Data;
-
-@Data
-public class ProyectoEstudianteDto {
-
-}
