--- src/main/java/ues/occ/proyeccion/social/ws/app/dto/StatusEnum.java
@@ -1,9 +1,0 @@
-package ues.occ.proyeccion.social.ws.app.dto;
-
-public enum StatusEnum {
-	Pendiente,
-	En_Proceso,
-	Completado,
-	Rechazado,
-	Retiro
-}
