+++ src/main/java/ues/occ/proyeccion/social/ws/app/model/PersonalExternoDto.java
@@ -1,4 +1,4 @@
-package ues.occ.proyeccion.social.ws.app.dto;
+package ues.occ.proyeccion.social.ws.app.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
