+++ src/main/java/ues/occ/proyeccion/social/ws/app/model/StatusEnum.java
@@ -1,0 +1,9 @@
+package ues.occ.proyeccion.social.ws.app.model;
+
+public enum StatusEnum {
+	PENDIENTE,
+	EN_PROCESO,
+	COMPLETADO,
+	RECHAZADO,
+	RETIRO
+}
