+++ src/main/java/ues/occ/proyeccion/social/ws/app/service/DocumentoService.java
@@ -3,7 +3,7 @@ package ues.occ.proyeccion.social.ws.app.service;
 import org.springframework.http.ResponseEntity;
 
 import ues.occ.proyeccion.social.ws.app.dao.ServiceResponse;
-import ues.occ.proyeccion.social.ws.app.dto.DocumentoRequest;
+import ues.occ.proyeccion.social.ws.app.model.DocumentoRequest;
 
 public interface DocumentoService {
