+++ src/test/java/ues/occ/proyeccion/social/ws/app/service/ProyectoServiceImplTest.java
@@ -392,7 +392,7 @@ class ProyectoServiceImplTest {
 
         var carnets = List.of(carnet);
 
-        ProyectoCreationDTO proyectoCreationDTO = new ProyectoCreationDTO("Project", 150, true, 1, carnets, new StatusDTO(1, "DummyStatus", "DummyDesc"));
+        ProyectoCreationDTO proyectoCreationDTO = new ProyectoCreationDTO("Project", 150, true, 1, carnets);
         Proyecto resultProject = this.proyectoMapper.proyectoCreationDTOToProyecto(proyectoCreationDTO);
         resultProject.setTutor(personal);
 
@@ -474,7 +474,7 @@ class ProyectoServiceImplTest {
         proyecto.setFechaModificacion(LocalDateTime.now());
         proyecto.setStatus(new Status(1, "dummyValue", "dummyValue"));
 
-        var creationDto = new ProyectoCreationDTO("Test2", 350, true, 10, List.of("ab12345"), new StatusDTO());
+        var creationDto = new ProyectoCreationDTO("Test2", 350, true, 10, List.of("ab12345"));
 
         ArgumentCaptor<Integer> idCaptor = ArgumentCaptor.forClass(Integer.class);
         ArgumentCaptor<Integer> idPersonal = ArgumentCaptor.forClass(Integer.class);
